+++ example/lib/bitmap_page.dart
@@ -1,10 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:arabic_text_justification/arabic_text_justification.dart';
 
-import 'common.dart';
+import 'constants/constants.dart';
 import 'main.dart';
-
-const _verseMarker = '۝';
+import 'mixins/ayah_selection_mixin.dart';
+import 'mixins/debounced_slider_mixin.dart';
+import 'widgets/scrollable_page.dart';
+import 'widgets/slider_header.dart';
 
 class BitmapPage extends StatefulWidget {
   const BitmapPage({super.key});
@@ -14,38 +16,37 @@ class BitmapPage extends StatefulWidget {
 }
 
 class _BitmapPageState extends State<BitmapPage>
-    with AyahSelectionMixin<BitmapPage>, FontSizeMixin<BitmapPage> {
+    with AyahSelectionMixin<BitmapPage>, DebouncedSliderMixin<BitmapPage> {
+  @override
+  double get initialSliderValue => 24;
+
   @override
   Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.end,
-      children: [
-        FontSizeHeader(
-          fontSize: fontSize,
-          onChanged: onFontSizeChanged,
-          selectedAyah: selectedAyah,
-          selectedWord: selectedWord,
-        ),
-        Expanded(
-          child: SingleChildScrollView(
-            padding: const EdgeInsets.symmetric(horizontal: 8),
-            child: Column(
-              children: [
-                for (int i = 0; i < page3Lines.length; i++)
-                  JustifiedArabicBitmapLine(
-                    words: page3Lines[i].words,
-                    justify: page3Lines[i].justify,
-                    fontSize: renderedFontSize,
-                    verseMarker: _verseMarker,
-                    highlightedWordIndices: highlightsFor(i),
-                    onWordTap: (idx, w) => onWordTap(i, idx, w),
-                    onMarkerTap: (idx, w) => onMarkerTap(i, idx, w),
-                  ),
-              ],
+    return ScrollablePage(
+      padding: const EdgeInsets.symmetric(horizontal: 8),
+      header: SliderHeader(
+        label: 'Font size',
+        value: sliderValue,
+        min: 8,
+        max: 64,
+        onChanged: onSliderChanged,
+        selectedAyah: selectedAyah,
+        selectedWord: selectedWord,
+      ),
+      child: Column(
+        children: [
+          for (int i = 0; i < page3Lines.length; i++)
+            JustifiedArabicBitmapLine(
+              words: page3Lines[i].words,
+              justify: page3Lines[i].justify,
+              fontSize: renderedValue,
+              marker: ayahMarker,
+              highlightedWordIndices: highlightsFor(i),
+              onWordTap: (idx, w) => onWordTap(i, idx, w),
+              onMarkerTap: (idx, w) => onMarkerTap(i, idx, w),
             ),
-          ),
-        ),
-      ],
+        ],
+      ),
     );
   }
 }