+++ example/lib/constants/constants.dart
@@ -1,0 +1,3 @@
+const ayahMarker = '۝';
+
+const wordDuration = Duration(milliseconds: 500);