+++ lib/arabic_text_justification.dart
@@ -1,3 +1,4 @@
+export 'src/models/word_color_span.dart';
 export 'src/models/word_progress.dart';
 export 'src/widgets/justified_arabic_bitmap_line.dart';
 export 'src/widgets/justified_arabic_line.dart';